feat(zsh): update to omz/powerlevel10k
I did this because I was tired of dealing with all of the other 'plugin'
managers. I seemed to have run into issues with just about every other
one. So far, out of the box, omz appears to have much of what I was
looking for. Time will tell as far as speed and efficiency go.